Let the post-response evaluation answer nothing at all
